restore folder check for scripts in clean

+++ tasks/clean.js
@@ -24,9 +24,11 @@ rmdir(path.join(__dirname, "..", "src", "transpiled"));
 
 // We don't remove Scripts, but instead remove child directories of Scripts
 const scriptsFolder = path.join(__dirname, "..", "Scripts");
-fs.readdirSync(scriptsFolder).forEach((file, index) => {
-    const subpath = path.join(scriptsFolder, file);
-    if (fs.lstatSync(subpath).isDirectory()) {
-        rmdir(subpath);
-    }
-});
+if (fs.existsSync(scriptsFolder)) {
+    fs.readdirSync(scriptsFolder).forEach((file, index) => {
+        const subpath = path.join(scriptsFolder, file);
+        if (fs.lstatSync(subpath).isDirectory()) {
+            rmdir(subpath);
+        }
+    });
+}
